convert `WeekDay.js` to a functional component
